--- frontend/src/component/Footer.js
@@ -1,8 +1,0 @@
-import React from "react";
-import { Container, Grid, Button, Typography } from "@material-ui/core";
-
-const Footer = () => {
-  return <Container>Copyright @ KwokkieShop</Container>;
-};
-
-export default Footer;
