+++ frontend/src/components/Footer.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { Container } from "@material-ui/core";
+
+const Footer = () => {
+    return <Container>Copyright @ KwokkieShop</Container>;
+};
+
+export default Footer;
